refactor(): put github client inside repo analyzer namespace
